+++ src/bda/plone/orders/common.py
@@ -176,6 +176,8 @@ class BookingsCatalogFactory(object):
 
     def __call__(self, context=None):
         catalog = Catalog()
+        email_indexer = NodeAttributeIndexer('email')
+        catalog[u'email'] = CatalogFieldIndex(email_indexer)
         uid_indexer = NodeAttributeIndexer('uid')
         catalog[u'uid'] = CatalogFieldIndex(uid_indexer)
         buyable_uid_indexer = NodeAttributeIndexer('buyable_uid')
@@ -196,6 +198,10 @@ class BookingsCatalogFactory(object):
         catalog[u'state'] = CatalogFieldIndex(state_indexer)
         salaried_indexer = NodeAttributeIndexer('salaried')
         catalog[u'salaried'] = CatalogFieldIndex(salaried_indexer)
+        search_attributes = ['email',
+                             'title']
+        text_indexer = NodeTextIndexer(search_attributes)
+        catalog[u'text'] = CatalogTextIndex(text_indexer)
         return catalog
 
 
@@ -371,6 +377,7 @@ class OrderCheckoutAdapter(CheckoutAdapter):
         vendor = acquire_vendor_or_shop_root(buyable)
 
         booking = OOBTNode()
+        booking.attrs['email'] = order.attrs['personal_data.email']
         booking.attrs['uid'] = uuid.uuid4()
         booking.attrs['buyable_uid'] = uid
         booking.attrs['buyable_count'] = count