+++ CHANGES.rst
@@ -5,9 +5,16 @@ Changelog
 0.9.dev0
 --------
 
-- Major cleanup - code-analysis integrated, travis ci and moved IBuyable here to
-  avoid circular dependency.
-  [bennyboy]
+- Add two datatable views, in which bookings are displayed and can be grouped
+  by the buyers email adress or the buyable uid. Both views support daterange
+  filtering and text index support. The ``Bookings`` view gets called on the
+  portal root and the ``Bookings in Context`` returns all bookings data on
+  the corresponding context it is called.
+  [benniboy]
+
+- Major cleanup - code-analysis integrated, travis ci and moved IBuyable from
+  bda.plone.shop to this package to avoid circular dependencies.
+  [benniboy]
 
 - Dont depend on implemented interfaces ITrading and IShippingItem.
   see https://github.com/bluedynamics/bda.plone.shop/issues/31